src/test/java/uk/gov/hmcts/reform/divorce/orchestration/controller/BulkCaseControllerTest.java
package uk.gov.hmcts.reform.divorce.orchestration.controller;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
import uk.gov.hmcts.reform.divorce.orchestration.service.CaseOrchestrationService;

import java.util.Collections;
import java.util.Map;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.when;

@RunWith(MockitoJUnitRunner.class)
public class BulkCaseControllerTest {

    @Mock
    private CaseOrchestrationService caseOrchestrationService;

    @InjectMocks
    private BulkCaseController classUnderTest;


    @Test
    public void whenSearchCases_thenReturnExpectedResponse() throws WorkflowException {
        Map<String, Object> expected = Collections.emptyMap();

        when(caseOrchestrationService.generateBulkCaseForListing()).thenReturn(expected);

        ResponseEntity<Map<String, Object>> response = classUnderTest.processCasesReadyForListing();

        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertEquals(expected, response.getBody());

    }
}
